+++ lib/src/utils/async_state.dart
@@ -1,43 +1,25 @@
 import 'package:async_notifier/async_notifier.dart';
 import 'package:flutter/material.dart';
 
-import '../widgets/async/async.dart';
-
 /// A [State] with an [AsyncNotifier].
 abstract class AsyncState<T extends StatefulWidget, Data> extends State<T> {
   /// The [AsyncNotifier] of this state.
-  late final async = AsyncNotifier.late<Data>(
-    value: initialData,
-    onError: onError,
-    onData: onData,
-  );
+  @protected
+  late final async = AsyncNotifier<Data>(data: initialData);
 
   /// The initial data of this state.
   Data? get initialData => null;
 
-  /// Called whenever [AsyncSnapshot] resolves with error.
-  @protected
-  @mustCallSuper
-  void onError(Object error, StackTrace stackTrace) {
-    Async.errorLogger(error, stackTrace);
-  }
-
-  /// Called when [AsyncSnapshot] resolves with new data.
-  @protected
-  @mustCallSuper
-  void onData(Data data) {}
+  /// The [AsyncSnapshot] of this state.
+  AsyncSnapshot<Data> get snapshot => async.snapshot;
 
   /// Called whenever [AsyncSnapshot] changes.
   void onSnapshot(AsyncSnapshot<Data> snapshot) {}
 
-  /// Performs [AsyncNotifier.cancel] on current async computation.
-  @mustCallSuper
-  void cancel() => async.cancel();
-
   @override
   void initState() {
     super.initState();
-    async.addListener(() => setState(() => onSnapshot(async.snapshot)));
+    async.addListener(() => setState(() => onSnapshot(snapshot)));
   }
 
   @override